feat: Button does not accept to props by default
