+++ packages/react-admin/package.json
@@ -39,7 +39,7 @@
         "@emotion/react": "^11.4.1",
         "@emotion/styled": "^11.3.0",
         "@mui/icons-material": "^5.0.1",
-        "@mui/material": "^5.0.2",
+        "@mui/material": "^5.14.10",
         "history": "^5.1.0",
         "ra-core": "^4.14.1",
         "ra-i18n-polyglot": "^4.14.1",
